Moved MaxQueue class to utils.py
